Revert "Finishing touches"
This reverts commit d8ff6f39cec6a62b75b29bc28b8810e4964b43ea.
Reverting README.md changes

+++ Cinema-back-end/src/main/java/com/app/cinema/Showing/Showing.java
@@ -1,6 +1,5 @@
 package com.app.cinema.Showing;
 
-import java.math.BigDecimal;
 import java.sql.Timestamp;
 
 public record Showing(
@@ -14,7 +13,7 @@ public record Showing(
         String room,
         Integer length_minutes,
         String language,
-        BigDecimal price,
+        Float price,
         String[] genres
 ) {
 }
