+++ Cinema-back-end/src/main/resources/db/migration/6__InitShowing.sql
@@ -1,7 +1,7 @@
 CREATE TABLE showing (
    id BIGSERIAL PRIMARY KEY,
    start TIMESTAMP NOT NULL,
-   price DECIMAL NOT NULL,
+   price FLOAT NOT NULL,
    room TEXT NOT NULL,
    movie BIGINT REFERENCES movie (id),
    cinema BIGINT REFERENCES cinema (id)
